+++ src/features/AddQuizSource/ui/AddQuizSource.tsx
@@ -18,7 +18,7 @@ export const AddQuizSource = ({ selectedQuiz, toggleMusicModal, toggleYoutubeMod
             <iframe 
               width="350" 
               height="250" 
-              src={selectedQuiz?.source}ß
+              src={selectedQuiz?.source}
               title="YouTube video player" 
               frameborder="0" 
               allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share"
