+++ src/features/AddQuizType/index.ts
@@ -1,0 +1,1 @@
+export { AddQuizType } from './ui/AddQuizType';
