+++ src/features/EditQuizSetting/ui/EditQuizSetting.tsx
@@ -5,7 +5,7 @@ const Wrapper = styled.div`
   height: 200px;
   background-color: white;
   filter: drop-shadow(0 0 10px rgba(0, 0, 0, 0.25));
-  margin: 12px 0px
+  margin: 12px 0;
 `
 
 const SettingInfo = styled.div`
@@ -31,39 +31,40 @@ const SettingValue = styled.div<{ selected: boolean }>`
   flex-direction: row;
   justify-content: center;
   align-items: center;
-  background-color:${(props) => (props.selected ? "#FD7400" : "white")};
-  color:${(props) => (props.selected ? "white" : "black")};
-  &:hover{
+  background-color: ${(props) => (props.selected ? "#FD7400" : "white")};
+  color: ${(props) => (props.selected ? "white" : "black")};
+  &:hover {
     background-color: #FD7400;
     cursor: pointer;
     color: white;
   }
-  border-top: 0.1px solid #E0E0E0;//TODO: 추후에 변경
+  border-top: 0.1px solid #E0E0E0; //TODO: 추후에 변경
 `
 
 type EditQuizSettingProps = {
   EditTitle: string;
   EditOption: any[];
-  Value:any;
-  HandleEdit?: (value:any) => void;
+  Value: any;
+  HandleEdit?: (value: any) => void;
 }
 
-export const EditQuizSetting = ({EditTitle, EditOption, Value, HandleEdit}:EditQuizSettingProps) =>{
-  return(
+export const EditQuizSetting = ({ EditTitle, EditOption, Value, HandleEdit }: EditQuizSettingProps) => {
+  const EditOptionByKorean = ["객관식", "주관식", "OX퀴즈"];
+
+  return (
     <Wrapper>
       <SettingInfo>{EditTitle}</SettingInfo>
       <SettingOption>
         {EditOption.map((option, index) => (
-          <SettingValue 
-            onClick={()=>HandleEdit(option)} 
+          <SettingValue
+            onClick={() => HandleEdit(option)}
             key={index}
             selected={option === Value}
-          >{option}
+          >
+            {EditTitle === "문제유형" ? EditOptionByKorean[index] : option}
           </SettingValue>
-          ))}
+        ))}
       </SettingOption>
     </Wrapper>
   )
 }
-
-//TODO: 문제유형 벨류와 따라서 보여주는 값 다르게 하가
